LayoutTests/platform/blackberry/fast/js/create-lots-of-workers-expected.txt

CONSOLE MESSAGE: line 14: INVALID_STATE_ERR: DOM Exception 11: An attempt was made to use an object that is not, or is no longer, usable.
FAIL: Timed out waiting for notifyDone to be called
